refactor(domain): move formatMonthDay to the shared format module
Lives alongside the other French date formatters (`formatShortDate`,
`formatLongDate`, `formatShortDateTime`) rather than as a module-local
helper under `admin/stats`. Reachable via the `~/modules/domain` barrel
for any future year-agnostic day-of-year renderer.
