+++ packages/app/src/modules/admin/stats/CampaignProgressionTable.tsx
@@ -1,4 +1,5 @@
-import { formatMonthDay } from "./format";
+import { formatMonthDay } from "~/modules/domain";
+
 import type { CampaignProgressionSeries } from "./types";
 
 type Props = {
