--- packages/app/src/modules/admin/stats/format.ts
@@ -1,8 +1,0 @@
-/**
- * Convert a `MM-DD` key (shared across years on the progression axis) to
- * the French display form `DD/MM`.
- */
-export function formatMonthDay(monthDay: string): string {
-	const [month, day] = monthDay.split("-");
-	return `${day}/${month}`;
-}
